requirements.txt
opencv-python-headless>=4.12.0.88
tensorflow>=2.20.0
Pillow>=11.0.0
numpy>=1.24.0
flask>=3.0.0
